Create 10.11 LAB Input errors
Write a program that takes in three integers as inputs and outputs the largest value. Use a try block to perform all the statements. Use a catch block to catch any NoSuchElementException caused by missing inputs. Then output the number of inputs read and the largest value, or output "No max" if no inputs are read.